Rework names to be called Command instead.
