+++ Gemfile
@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in simple_command.gemspec
+# Specify your gem's dependencies in command.gemspec
 gemspec
 
 group :development do
