+++ command.gemspec
@@ -1,17 +1,17 @@
 # coding: utf-8
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'simple_command/version'
+require 'command/version'
 
 Gem::Specification.new do |s|
   s.required_ruby_version = '>= 2.0'
-  s.name          = 'simple_command'
-  s.version       = SimpleCommand::VERSION
-  s.authors       = ['Andrea Pavoni']
-  s.email         = ['andrea.pavoni@gmail.com']
+  s.name          = 'command'
+  s.version       = Command::VERSION
+  s.authors       = ['Andrea Pavoni', 'Guillaume Charneau', 'Jérémie Bonal']
+  s.email         = ['andrea.pavoni@gmail.com', 'guillaume.charneau@swile.co', 'jeremie.bonal@swile.co']
   s.summary       = 'Easy way to build and manage commands (service objects)'
   s.description   = 'Easy way to build and manage commands (service objects)'
-  s.homepage      = 'http://github.com/nebulab/simple_command'
+  s.homepage      = 'http://github.com/TheMenu/command'
   s.license       = 'MIT'
 
   s.files         = `git ls-files -z`.split("\x0")
