+++ spec/command/error_handling_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe SimpleCommand::ErrorHandling do
+describe Command::ErrorHandling do
   before do
     instance
 
@@ -9,7 +9,7 @@ describe SimpleCommand::ErrorHandling do
 
   let(:klass) do
     prepend_in = Class.new
-    prepend_in.prepend(SimpleCommand) # To provide contract ErrorHandling relies on
+    prepend_in.prepend(Command::SimpleCommand) # To provide contract ErrorHandling relies on
     prepend_in.prepend(described_class)
   end
