+++ spec/factories/success_command.rb
@@ -1,5 +1,5 @@
 class SuccessCommand
-  prepend SimpleCommand
+  prepend Command::SimpleCommand
 
   def initialize(input)
     @input = input
