+++ spec/spec_helper.rb
@@ -6,7 +6,7 @@ end
 
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 
-require 'simple_command'
+require 'command'
 
 Dir[File.join(File.dirname(__FILE__), 'factories', '**/*.rb')].each do |factory|
   require factory
